[Design] : UI 1차 작업
