+++ app/src/main/java/com/example/kakaoimglibrary/bookmark/BookmarkViewModel.kt
@@ -1,0 +1,4 @@
+package com.example.kakaoimglibrary.bookmark
+
+class BookmarkViewModel {
+}
