+++ app/src/main/java/com/example/kakaoimglibrary/main/SharedViewModel.kt
@@ -1,0 +1,4 @@
+package com.example.kakaoimglibrary.main
+
+class SharedViewModel {
+}
