+++ app/src/main/java/com/example/kakaoimglibrary/search/SearchFragment.kt
@@ -1,18 +1,17 @@
-package com.example.kakaoimglibrary
+package com.example.kakaoimglibrary.search
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import com.example.kakaoimglibrary.R
 
-
-class BookmarkFragment : Fragment() {
+class SearchFragment : Fragment() {
 
     companion object{
-        fun newInstance() = BookmarkFragment()
+        fun newInstance() = SearchFragment()
     }
-
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
@@ -24,7 +23,7 @@ class BookmarkFragment : Fragment() {
         savedInstanceState: Bundle?
     ): View? {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_bookmark, container, false)
+        return inflater.inflate(R.layout.fragment_search, container, false)
     }
 
 }
