+++ app/src/main/java/com/example/kakaoimglibrary/search/SearchViewModel.kt
@@ -1,0 +1,4 @@
+package com.example.kakaoimglibrary.search
+
+class SearchViewModel {
+}
