+++ app/src/main/res/layout/fragment_bookmark.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".BookmarkFragment">
+    tools:context=".bookmark.BookmarkFragment">
 
     <TextView
         android:layout_width="match_parent"
