+++ app/src/main/res/values/strings.xml
@@ -4,5 +4,7 @@
     <string name="hello_blank_fragment">Hello blank fragment</string>
     <string name="main_tab_search_title">Home</string>
     <string name="main_tab_bookmark_title">BookMark</string>
+    <string name="test_title">[image] 테스트용 이미지 입니다.</string>
+    <string name="test_time">2023.09.19 10:06:00</string>
 
 </resources>
